+++ kinto/core/resource/__init__.py
@@ -303,7 +303,7 @@ class UserResource(object):
             :meth:`kinto.core.resource.UserResource.process_record`
         """
         existing = None
-        new_record = self.request.validated['data']
+        new_record = self.request.validated.get('data', {})
         try:
             id_field = self.model.id_field
             # Since ``id`` does not belong to schema, look up in body.
@@ -413,7 +413,8 @@ class UserResource(object):
             if existing:
                 self._raise_412_if_modified(existing)
 
-        post_record = self.request.validated['data']
+        # If `data` is not provided, use existing record (or empty if creation)
+        post_record = self.request.validated.get('data', existing) or {}
 
         record_id = post_record.setdefault(id_field, self.record_id)
         self._raise_400_if_id_mismatch(record_id, self.record_id)