+++ CHANGELOG.rst
@@ -14,6 +14,10 @@ This document describes changes between each past release.
   it by ``%(http_port)s`` or regenerate a new configuration file with ``kinto init``.
 - Add support for ``pool_timeout`` option in Redis backend (fixes #620)
 
+**Bug fixes**
+
+- Fix loss of data attributes when permissions are replaced with ``PUT`` (fixes #601)
+
 
 3.1.0 (2016-05-24)
 ==================